requirements.txt
asttokens==3.0.0
azure-ai-translation-document==1.1.0
azure-ai-translation-text==1.0.1
azure-core==1.33.0
azure-identity==1.21.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.14
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
msal==1.32.3
msal-extensions==1.3.1
mypy==1.15.0
mypy_extensions==1.1.0
nest_asyncio==1.6.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pip==25.0.1
platformdirs==4.3.7
polars==1.27.1
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyzmq==26.4.0
requests==2.32.3
setuptools==79.0.1
six==1.17.0
stack_data==0.6.3
tornado==6.4.2
traitlets==5.14.3
types-requests==2.32.0.20250328
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uuid==1.30
wcwidth==0.2.13
wheel==0.45.1
zipp==3.21.0
